Rename meta-processor package && support hive compatible
