+++ odps-data-carrier/bin/meta-processor
@@ -1,3 +1,3 @@
 #!/bin/bash
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-java -cp "$DIR/../libs/*" com.aliyun.odps.datacarrier.odps.datacarrier.MetaProcessor $@
+java -cp "$DIR/../libs/*" com.aliyun.odps.datacarrier.metaprocessor.MetaProcessor $@
