+++ odps-data-carrier/meta-processor/src/main/java/com/aliyun/odps/datacarrier/metaprocessor/TypeTransformResult.java
@@ -1,4 +1,4 @@
-package com.aliyun.odps.datacarrier.odps.datacarrier;
+package com.aliyun.odps.datacarrier.metaprocessor;
 
 import com.aliyun.odps.datacarrier.commons.Constants.DATASOURCE_TYPE;
 import com.aliyun.odps.datacarrier.commons.risk.Risk;
