+++ odps-data-carrier/meta-processor/src/main/java/com/aliyun/odps/datacarrier/metaprocessor/report/SummaryDivision.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package com.aliyun.odps.datacarrier.odps.datacarrier.report;
+package com.aliyun.odps.datacarrier.metaprocessor.report;
 
 import htmlflow.HtmlView;
 import java.util.HashMap;
